Update contact information
The email addresses of the existing contacts are being changed from .com
to .org as this will serve as the primary email domain (at least until
we decide to switch to bigapplepy.org for email). Also, the contact
information for the officers not already listed is being added to
recognize their contributions.