Switch to Jekyll
This site is hopefully going to switch from a single HTML file to
something more useful. Rather than creating each new page by hand (and
updating any others that need to link to it), a static site generator is
going to be used. For the time being, that's going to be Jekyll.

While this implementation doesn't really take any advantage of Jekyll,
it's goal is just to get the current version working through it.